Tried to fix timing issues; CSD using seconds, game using ticks

+++ core/src/com/combatcube/pianoshooter/Shooter.java
@@ -24,7 +24,7 @@ public class Shooter {
         updateTouchBox();
     }
 
-    public void moveY(long tick) {
+    public void moveY(float tick) {
         rect.setY(tick);
         updateTouchBox();
     }
